+++ tests/testthat/test-tweakPedigree.R
@@ -18,7 +18,8 @@ test_that("makeTwins - Twins specified by IDs", {
     momID = c(NA, NA, 2, 2),
     spID = c(NA, NA, NA, NA),
     sex = c("M", "F", "M", "F"),
-    MZtwin = c(2, 1, NA, NA)
+    twinID = c(2, 1, NA, NA),
+    zygosity = c("MZ", "MZ", NA, NA)
   )
   result <- makeTwins(ped, ID_twin1 = 1, ID_twin2 = 2)
 
@@ -29,6 +30,8 @@ test_that("makeTwins - Twins specified by IDs", {
 
   result <- makeTwins(ped, ID_twin1 = 1, ID_twin2 = 2, verbose = TRUE)
   expect_equal(result, expected_result)
+
+ # hp <- makeTwins(potter, ID_twin1 = 12, ID_twin2 = 13, verbose = TRUE)
 })
 
 test_that("makeTwins - Twins specified by generation", {
@@ -41,21 +44,21 @@ test_that("makeTwins - Twins specified by generation", {
   ped <- simulatePedigree(kpc = kpc, Ngen = Ngen, sexR = sexR, marR = marR)
   #
   result <- makeTwins(ped, gen_twin = gen_twin)
-  expect_equal(names(result), c("famID", "ID", "gen", "dadID", "momID", "spID", "sex", "MZtwin"))
+  expect_equal(names(result), c("famID", "ID", "gen", "dadID", "momID", "spID", "sex", "twinID", "zygosity"))
   # do we have the same people?
   expect_equal(result$ID, ped$ID)
   # did it make one pair of twins?
-  expect_equal(sum(!is.na(result$MZtwin)), 2)
+  expect_equal(sum(!is.na(result$twinID)), 2)
   # did it make the pair in the correct generation?
-  expect_equal(mean(result$gen[!is.na(result$MZtwin)]), gen_twin)
+  expect_equal(mean(result$gen[!is.na(result$twinID)]), gen_twin)
   # are they the same sex?
-  expect_equal(length(unique(result$sex[!is.na(result$MZtwin)])), 1)
+  expect_equal(length(unique(result$sex[!is.na(result$twinID)])), 1)
   # are they from the same family?
-  expect_equal(length(unique(result$fam[!is.na(result$MZtwin)])), 1)
+  expect_equal(length(unique(result$fam[!is.na(result$twinID)])), 1)
   # do they have the same mom?
-  expect_equal(length(unique(result$momID[!is.na(result$MZtwin)])), 1)
+  expect_equal(length(unique(result$momID[!is.na(result$twinID)])), 1)
   # do they have the same dad?
-  expect_equal(length(unique(result$dadID[!is.na(result$MZtwin)])), 1)
+  expect_equal(length(unique(result$dadID[!is.na(result$twinID)])), 1)
 })
 
 # Test for makeInbreeding function