+++ NEWS.md
@@ -1,5 +1,6 @@
 # BGmisc 1.4.2
 * Added twinIDs for potter and asoiaf pedigrees
+* Added twinID to simulatePedigree function, and extended to include MZ, DZ, and SS twins.
 
 # BGmisc 1.4.1
 * replaced print with message in all functions